+++ CHANGES.md
@@ -3,6 +3,7 @@
 * add filter `wpbt_hide_report_a_bug`
 * update if `Report a Bug` plugin active
 * link to settings if `Report a Bug` is hidden
+* update for latest WP API responses, thanks @dd32
 
 #### 3.5.0 / 2023-06-30
 * update dashboard widget for MarComm publishing of posts